Second kid of the day
So, this one is basically the opposite.
add in the values and whatsoever, it tells you the colors of the
resistor you're looking for
Took some time
Also, I was lazy to make both of them foolproof, so just try and stick
to using them as a normal human being, and not some up-jumped little
circus monkey that's just found a computer.